pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prism-signal"
version = "1.0.0"
description = "Signal analysis engine: primitives, engines, and pipelines for time-series feature extraction"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Jason Rudder"}
]
keywords = ["signal-processing", "time-series", "eigenvalues", "entropy", "complexity", "feature-extraction"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Scientific/Engineering :: Mathematics"
]
requires-python = ">=3.9"

dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "polars>=0.20.0",
    "joblib>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "mypy"
]

dynamics = [
    "nolds>=0.5.0",
]

ml = [
    "scikit-learn>=1.0.0",
]

all = [
    "prism-signal[dynamics,ml]"
]

[project.urls]
Homepage = "https://github.com/prism-engines/prism"
Repository = "https://github.com/prism-engines/prism"
Documentation = "https://prism-engines.github.io/prism"
"Bug Reports" = "https://github.com/prism-engines/prism/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["prism*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--cov=prism --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
